backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
anthropic==0.52.0
google-genai==1.14.0
pydantic==2.11.1
pydantic-settings==2.7.1
python-dotenv==1.1.0
aiofiles==24.1.0
pytest==8.3.4
httpx==0.28.1
